# backend/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/v1/auth/session": {
            "get": {
                "operationId": "auth.session",
                "description": "Returns the currently authenticated user.",
                "summary": "GET /api/v1/auth/session  (auth:sanctum)",
                "tags": [
                    "Auth"
                ],
                "responses": {
                    "200": {
                        "description": "setStatusCode(200): a JsonResource whose model wasRecentlyCreated (first OIDC\nlogin creates the account) would otherwise auto-return 201; auth is not a create.\n\n\n\n`AccountSessionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/AccountSessionResource"
                                        }
                                    },
                                    "required": [
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountSessionResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "email_verified": {
                        "type": "string"
                    },
                    "startup_gate_subject_id": {
                        "type": "string"
                    },
                    "linked_providers": {
                        "type": "string"
                    },
                    "has_password": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "id",
                    "email",
                    "display_name",
                    "email_verified",
                    "startup_gate_subject_id",
                    "linked_providers",
                    "has_password"
                ],
                "title": "AccountSessionResource"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}